Restructure packages for framework separation
