+++ packages/react/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "el-form",
+  "name": "@el-form/react",
   "version": "0.1.0",
   "main": "dist/index.js",
   "types": "dist/index.d.ts",
@@ -16,6 +16,7 @@
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "zod": "^3.22.4",
-    "@el-form/react": "workspace:*"
+    "@el-form/auto-form": "workspace:*",
+    "@el-form/use-form": "workspace:*"
   }
 }
