+++ src/index.ts
@@ -1,12 +1,1 @@
-export { useForm, Form } from "../packages/use-form/src";
-export type { FormState, UseFormOptions, UseFormReturn } from "../packages/use-form/src/types";
-export { AutoForm } from "../packages/auto-form/src";
-export type {
-  AutoFormProps,
-  AutoFormFieldConfig,
-  AutoFormFieldProps,
-  AutoFormErrorProps,
-  GridColumns,
-  FieldConfig,
-} from "../packages/auto-form/src/types";
-export { parseZodErrors } from "../packages/use-form/src/utils";
+export * from "../packages/el-form/src";
